remove all the specializations and just use value_of() and coercing function. Fixup tests to check for types and move var versions over to rev
