+++ stan/math/rev/fun.hpp
@@ -16,6 +16,8 @@
 #include <stan/math/rev/fun/acos.hpp>
 #include <stan/math/rev/fun/acosh.hpp>
 #include <stan/math/rev/fun/as_bool.hpp>
+#include <stan/math/rev/fun/as_array_or_scalar.hpp>
+#include <stan/math/rev/fun/as_column_vector_or_scalar.hpp>
 #include <stan/math/rev/fun/arg.hpp>
 #include <stan/math/rev/fun/asin.hpp>
 #include <stan/math/rev/fun/asinh.hpp>
