+++ test/unit/math/prim/fun/as_value_column_vector_or_scalar_test.cpp
@@ -20,7 +20,8 @@ TEST(MathMetaPrim, as_value_column_vector_or_scalar_std_vector_lvalue) {
     b[i] = a.coeff(i);
   }
   auto&& tmp = as_value_column_vector_or_scalar(b);
-  a[0] = tmp[0] = 12345;
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
+  a[0] = b[0] = 12345;
   Eigen::VectorXd res = tmp;
   EXPECT_MATRIX_EQ(res, a);
 }
@@ -34,6 +35,7 @@ TEST(MathMetaPrim, as_value_column_vector_or_scalar_std_vector_rvalue) {
     b[i] = a.coeff(i);
   }
   const auto& tmp = as_value_column_vector_or_scalar(std::move(b));
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
   b.assign(n, 0);  // overwrite the memory if b was not moved
   Eigen::VectorXd res = tmp;
   EXPECT_MATRIX_EQ(res, a);
@@ -44,6 +46,7 @@ TEST(MathMetaPrim, as_value_column_vector_or_scalar_const_row_vector_lvalue) {
   int n = 100;
   const Eigen::RowVectorXd a = Eigen::RowVectorXd::Random(n);
   auto&& tmp = as_value_column_vector_or_scalar(a);
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
   Eigen::VectorXd res = tmp;
   Eigen::VectorXd correct = a.transpose();
   EXPECT_MATRIX_EQ(res, correct);
@@ -54,6 +57,7 @@ TEST(MathMetaPrim, as_value_column_vector_or_scalar_row_vector_lvalue) {
   int n = 100;
   Eigen::RowVectorXd a = Eigen::RowVectorXd::Random(n);
   auto&& tmp = as_value_column_vector_or_scalar(a);
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
   tmp[0] = 1234;
   Eigen::VectorXd res = tmp;
   Eigen::VectorXd correct = a.transpose();
@@ -66,8 +70,9 @@ TEST(MathMetaPrim, as_value_column_vector_or_scalar_row_vector_rvalue) {
   Eigen::RowVectorXd a = Eigen::RowVectorXd::Random(n);
   Eigen::RowVectorXd b = a;
   const auto& tmp = as_value_column_vector_or_scalar(std::move(b));
-  Eigen::VectorXd res = tmp;
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
   b.setZero();  // overwrite the memory if b was not moved
+  Eigen::VectorXd res = tmp;
   Eigen::VectorXd correct = a.transpose();
   EXPECT_MATRIX_EQ(res, correct);
 }
