+++ test/unit/math/rev/fun/as_array_or_scalar_test.cpp
@@ -8,6 +8,7 @@ TEST(MathFunRev, as_array_or_scalar_var_value_matrix) {
   const Eigen::MatrixXd a_val = Eigen::MatrixXd::Random(n, n);
   stan::math::var_value<Eigen::MatrixXd> a(a_val);
   const auto& tmp = stan::math::as_array_or_scalar(a);
+
   EXPECT_MATRIX_EQ(tmp.val(), a_val);
 }
 
@@ -16,6 +17,7 @@ TEST(MathFunRev, as_array_or_scalar_var_value_vector) {
   const Eigen::VectorXd a_val = Eigen::VectorXd::Random(n);
   stan::math::var_value<Eigen::VectorXd> a(a_val);
   const auto& tmp = stan::math::as_array_or_scalar(a);
+  EXPECT_TRUE((stan::is_eigen_array<decltype(tmp.val())>::value));
   EXPECT_MATRIX_EQ(tmp.val(), a_val);
 }
 
@@ -24,5 +26,6 @@ TEST(MathFunRev, as_array_or_scalar_var_value_rowvector) {
   const Eigen::RowVectorXd a_val = Eigen::RowVectorXd::Random(n);
   stan::math::var_value<Eigen::RowVectorXd> a(a_val);
   const auto& tmp = stan::math::as_array_or_scalar(a);
+  EXPECT_TRUE((stan::is_eigen_array<decltype(tmp.val())>::value));
   EXPECT_MATRIX_EQ(tmp.val(), a_val);
 }
