+++ test/unit/math/rev/fun/as_value_column_vector_or_scalar_test.cpp
@@ -18,6 +18,7 @@ TEST(MathFunRev, as_value_column_vector_or_scalar_std_vector_lvalue) {
   }
   std::vector<stan::math::var> b(b_val.begin(), b_val.end());
   const auto& tmp = stan::math::as_value_column_vector_or_scalar(b);
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
   Eigen::VectorXd res = tmp;
   EXPECT_MATRIX_EQ(res, a_val);
 }
@@ -31,6 +32,7 @@ TEST(MathFunRev, as_value_column_vector_or_scalar_std_vector_rvalue) {
   }
   std::vector<stan::math::var> b(b_val.begin(), b_val.end());
   const auto& tmp = stan::math::as_value_column_vector_or_scalar(std::move(b));
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
   Eigen::VectorXd res = tmp;
   EXPECT_MATRIX_EQ(res, a_val);
 }
@@ -40,6 +42,7 @@ TEST(MathFunRev, as_value_column_vector_or_scalar_vector_lvalue) {
   Eigen::VectorXd a_val = Eigen::VectorXd::Random(n);
   Eigen::Matrix<stan::math::var, -1, 1> a(a_val);
   auto&& tmp = stan::math::as_value_column_vector_or_scalar(a);
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
   Eigen::VectorXd res = tmp;
   EXPECT_MATRIX_EQ(res, a_val);
 }
@@ -49,6 +52,7 @@ TEST(MathFunRev, as_value_column_vector_or_scalar_const_rowvector_lvalue) {
   const Eigen::RowVectorXd a_val = Eigen::RowVectorXd::Random(n);
   Eigen::Matrix<stan::math::var, 1, -1> a(a_val);
   auto&& tmp = stan::math::as_value_column_vector_or_scalar(a);
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
   Eigen::VectorXd res = tmp;
   EXPECT_MATRIX_EQ(res, a_val.transpose());
 }
@@ -59,6 +63,7 @@ TEST(MathFunRev, as_value_column_vector_or_scalar_rowvector_rvalue) {
   Eigen::Matrix<stan::math::var, 1, -1> a(a_val);
   auto b = a;
   const auto& tmp = stan::math::as_value_column_vector_or_scalar(std::move(b));
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
   Eigen::VectorXd res = tmp;
   EXPECT_MATRIX_EQ(res, a_val.transpose());
 }
@@ -68,6 +73,7 @@ TEST(MathFunRev, as_value_column_vector_or_scalar_var_value_vector) {
   const Eigen::VectorXd a_val = Eigen::VectorXd::Random(n);
   stan::math::var_value<Eigen::VectorXd> a(a_val);
   auto&& tmp = stan::math::as_value_column_vector_or_scalar(a);
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
   Eigen::VectorXd res = tmp;
   EXPECT_MATRIX_EQ(res, a_val);
 }
@@ -77,6 +83,7 @@ TEST(MathFunRev, as_value_column_vector_or_scalar_var_value_row_vector) {
   const Eigen::RowVectorXd a_val = Eigen::VectorXd::Random(n);
   stan::math::var_value<Eigen::RowVectorXd> a(a_val);
   auto&& tmp = stan::math::as_value_column_vector_or_scalar(a);
+  EXPECT_TRUE((stan::is_col_vector<decltype(tmp)>::value));
   Eigen::VectorXd res = tmp;
   EXPECT_MATRIX_EQ(res, a_val.transpose());
 }
